test: move palette and clipboard E2E helpers to the shared helpers module
